Source/mya/Source/Library/mya/src/main/java/com/philips/platform/mya/settings/SettingsModel.java
package com.philips.platform.mya.settings;


import com.philips.platform.mya.model.DataModel;

public class SettingsModel implements DataModel{

    private int itemCount=1;
    private String firstItem;
    private String secondItem;
    private String thirdItem;

    public int getItemCount() {
        return itemCount;
    }

    public void setItemCount(int itemCount) {
        this.itemCount = itemCount;
    }

    public String getFirstItem() {
        return firstItem;
    }

    public void setFirstItem(String firstItem) {
        this.firstItem = firstItem;
    }

    public String getSecondItem() {
        return secondItem;
    }

    public void setSecondItem(String secondItem) {
        this.secondItem = secondItem;
    }

    public String getThirdItem() {
        return thirdItem;
    }

    public void setThirdItem(String thirdItem) {
        this.thirdItem = thirdItem;
    }
}
